Drop APi server dependency (#132)
Coupling to the api server is expensive. It means we need to ship all our dependencies everywhere. It means we have to use github packages and manage another set of authentication tokens. And it means that we need to use the same versions of scala and libraries.

We also no longer need it. When definitions were stored by a schemaless service, sharing types meant avoiding errors. But if we store it in a relational database, schemas are enforced.